requirements.txt
anyio==3.7.1
attrs==22.1.0
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==38.0.2
cycler==0.11.0
dnspython==2.2.1
docopt==0.6.2
eventlet==0.33.1
exceptiongroup==1.0.4
fastapi==0.103.2
Flask==2.2.5
flit_core==3.6.0
fonttools==4.25.0
gevent==22.10.2
greenlet==3.0.3
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.4
jsonschema==4.17.3
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
networkx==2.6.3
numpy==1.21.5
packaging==22.0
Pillow==9.4.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
ply==3.11
pycparser==2.21
pydantic==1.10.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-engineio==3.14.2
python-socketio==4.6.0
requests==2.31.0
scipy==1.7.3
simple-websocket==1.0.0
sip==6.6.2
six==1.16.0
sniffio==1.2.0
starlette==0.27.0
toml==0.10.2
tornado==6.2
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
Werkzeug==2.2.3
wincertstore==0.2
wsproto==1.2.0
yarg==0.1.10
zipp==3.15.0
zope.event==5.0
zope.interface==6.4.post2
